requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
cloudpickle==2.2.1
tensorflow-cpu==2.16.1
pydantic==2.6.4
google-cloud-storage==2.16.0
